requirements.txt
Django==3.2.10
psycopg2==2.9.1
black==23.3.0
pylint==2.17.4
djangorestframework==3.14.0